+++ CHANGELOG.md
@@ -2,8 +2,13 @@
 
 ## Current Master
 
+## 2.9.0 2016-08-22
+
+- Added ShapeShift to the transaction history
+- Added affiliate key to Shapeshift requests
 - Added feature to reflect current conversion rates of current vault balance.
 - Changed transaction approval from notifications system to popup system.
+- Modify balance display logic.
 
 ## 2.8.0 2016-08-15